requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-import-export==2.5.0
django-queryset-csv==1.1.0
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0
